feat: extract WelcomeContent, remove example prompts
