+++ cmd/evener-hub/frontend/src/panes/welcome/welcome.module.css
@@ -10,11 +10,11 @@
   /* Issue #197: an unbreakable child (the resume-candidate button's
    * server-truncated-but-still-200-rune title, forced to white-space:nowrap
    * by button.module.css) sizes this flex column to its min-content width
-   * instead of the viewport's, which then propagates to .examples/.hints
-   * (their width:min(100%,36rem) resolves 100% against this inflated width)
-   * and is clipped mid-string by PaneScaffold's .body overflow-x:clip. Cap
-   * the column's width and let long words wrap so a wide child wraps instead
-   * of inflating the layout. */
+   * instead of the viewport's, which then propagates to .hints (its
+   * width:min(100%,36rem) resolves 100% against this inflated width) and is
+   * clipped mid-string by PaneScaffold's .body overflow-x:clip. Cap the
+   * column's width and let long words wrap so a wide child wraps instead of
+   * inflating the layout. */
   width: 100%;
   max-width: 36rem;
   overflow-wrap: break-word;
@@ -26,16 +26,13 @@
  * button.module.css (correct for most buttons - catastrophic here, where the
  * label is the widest thing in the column, not the narrowest). Override that
  * LOCALLY for buttons in this welcome .actions column so the long title wraps
- * instead of inflating the flex layout past the viewport - the same override
- * .examples button below applies for the identical reason. Wrapping reopens
- * the bug .button's nowrap exists to prevent (button.module.css pins
- * height: 32px on .md, so a wrapped label spills out of the box - 19px of
- * clipped text at 390px in the layoutguard probe), so the box also gets
- * height: auto with the desktop height kept as a floor. The phone tap floor
- * is re-declared in the media block because this rule's higher specificity
- * would otherwise beat button.module.css's own @media min-height. This rule
- * stays ABOVE .examples button so that rule's larger 44px floor still wins
- * for the suggestion buttons on desktop. */
+ * instead of inflating the flex layout past the viewport. Wrapping reopens
+ * the bug .button's nowrap exists to prevent (button.module.css pins height:
+ * 32px on .md, so a wrapped label spills out of the box - 19px of clipped
+ * text at 390px in the layoutguard probe), so the box also gets height: auto
+ * with the desktop height kept as a floor. The phone tap floor is re-declared
+ * in the media block because this rule's higher specificity would otherwise
+ * beat button.module.css's own @media min-height. */
 .actions button {
   white-space: normal;
   height: auto;
@@ -55,28 +52,6 @@
   font-family: var(--font-sans);
   font-size: var(--font-size-body);
   line-height: var(--line-height-body);
-}
-
-.examples {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  width: min(100%, 36rem);
-  gap: var(--space-2);
-  margin-top: var(--space-2);
-}
-
-.examplesLabel {
-  margin: 0;
-  color: var(--ink-mid);
-  font-family: var(--font-sans);
-  font-size: var(--font-size-ui);
-}
-
-.examples button {
-  min-height: 44px;
-  white-space: normal;
-  text-align: left;
 }
 
 /* T6: a quiet hint row teaching the chords, below the task suggestions -
